+++ AdminWebsite/AdminWebsite/Helper/SnakeCasePropertyNamingPolicy.cs
@@ -1,0 +1,17 @@
+using System.Linq;
+using System.Text.Json;
+
+namespace AdminWebsite.Helper
+{
+    public class SnakeCasePropertyNamingPolicy : JsonNamingPolicy
+    {
+        public override string ConvertName(string name)
+        {
+            return string.Concat(name.Select((character, index) =>
+                    index > 0 && char.IsUpper(character)
+                        ? "_" + character
+                        : character.ToString()))
+                .ToLower();
+        }
+    }
+}
